Move custom editor provider out of proposed to main api
For #77131
